+++ src/Service/VendorService/RbDigital/RbDigitalBooksVendorService.php
@@ -6,10 +6,8 @@
 
 namespace App\Service\VendorService\RbDigital;
 
-use App\Service\VendorService\ProgressBarTrait;
 use App\Service\VendorService\VendorServiceInterface;
 use App\Service\VendorService\VendorServiceTrait;
-use App\Utils\Message\VendorImportResultMessage;
 
 /**
  * Class RbDigitalBooksVendorService.
@@ -18,16 +16,7 @@ use App\Utils\Message\VendorImportResultMessage;
  */
 class RbDigitalBooksVendorService implements VendorServiceInterface
 {
-    use ProgressBarTrait;
     use VendorServiceTrait;
 
     protected const VENDOR_ID = 7;
-
-    /**
-     * {@inheritdoc}
-     */
-    public function load(): VendorImportResultMessage
-    {
-        return VendorImportResultMessage::error('Vendor deprecated');
-    }
 }
