+++ src/Service/VendorService/RbDigital/RbDigitalMagazinesVendorService.php
@@ -9,7 +9,6 @@ namespace App\Service\VendorService\RbDigital;
 use App\Service\VendorService\ProgressBarTrait;
 use App\Service\VendorService\VendorServiceInterface;
 use App\Service\VendorService\VendorServiceTrait;
-use App\Utils\Message\VendorImportResultMessage;
 
 /**
  * Class RbDigitalMagazinesVendorService.
@@ -22,12 +21,4 @@ class RbDigitalMagazinesVendorService implements VendorServiceInterface
     use VendorServiceTrait;
 
     protected const VENDOR_ID = 8;
-
-    /**
-     * {@inheritdoc}
-     */
-    public function load(): VendorImportResultMessage
-    {
-        return VendorImportResultMessage::error('Vendor deprecated');
-    }
 }
